feat: convert default value spec to form value
Currently, the OpenAPI support default value specification, although it's more intendened to be
treated as an optional field and is better handled on the server side. Still, there might be cases
where htis could be appropriate to be set as default in the client side, for example, having a
country form value set as `Singapore`, since most of our users are from `Singpoare`.